requirements.txt
requests==2.31.0
python-dotenv==0.21.1
python-telegram-bot==13.0